+++ src/controller_cli_callbacks.c
@@ -727,6 +727,65 @@ transaction_exist(clixon_handle h,
     return retval;
 }
 
+/*! Print warning for skipped devices in a completed transaction
+ *
+ * @param[in] h      Clixon handle
+ * @param[in] tidstr Transaction identifier string
+ * @retval    0      OK
+ * @retval   -1      Error
+ */
+static int
+transaction_print_skipped(clixon_handle h,
+                          char         *tidstr)
+{
+    int    retval = -1;
+    cxobj *xn = NULL;
+    cxobj *xerr;
+    cxobj *xdevices;
+    cxobj *xskip;
+    cvec  *nsc = NULL;
+    cbuf  *cb = NULL;
+    char  *name;
+    char  *reason;
+
+    if ((nsc = xml_nsctx_init("co", CONTROLLER_NAMESPACE)) == NULL)
+        goto done;
+    if ((cb = cbuf_new()) == NULL){
+        clixon_err(OE_PLUGIN, errno, "cbuf_new");
+        goto done;
+    }
+    cprintf(cb, "co:transactions/co:transaction[co:tid='%s']/co:devices", tidstr);
+    if (clicon_rpc_get(h, cbuf_get(cb), nsc, CONTENT_ALL, -1, "report-all", &xn) < 0)
+        goto done;
+    if ((xerr = xpath_first(xn, NULL, "/rpc-error")) != NULL){
+        clixon_err_netconf(h, OE_XML, 0, xerr, "Get transactions");
+        goto done;
+    }
+    xdevices = xpath_first(xn, nsc, "transactions/transaction/devices");
+    if (xdevices == NULL)
+        goto ok;
+    xskip = NULL;
+    while ((xskip = xml_child_each(xdevices, xskip, CX_ELMNT)) != NULL){
+        if (strcmp(xml_name(xskip), "skipped") != 0)
+            continue;
+        name = xml_find_body(xskip, "name");
+        reason = xml_find_body(xskip, "reason");
+        cligen_output(stderr, "Warning: device '%s' skipped (%s)\n",
+                      name ? name : "?",
+                      reason ? reason : "?");
+    }
+ ok:
+    retval = 0;
+  done:
+    if (cb)
+        cbuf_free(cb);
+    if (nsc)
+        cvec_free(nsc);
+    if (xn)
+        xml_free(xn);
+    return retval;
+}
+
 /*! Read(pull) the config of one or several devices.
  *
  * @param[in] h
@@ -815,8 +874,8 @@ cli_rpc_pull(clixon_handle h,
     if (ret == 1){
         if (transaction_notification_poll(h, tidstr, &result) < 0)
             goto done;
-        if (result == TR_SUCCESS)
-            cligen_output(stderr, "OK\n");
+        if (transaction_print_skipped(h, tidstr) < 0)
+            goto done;
     }
     retval = 0;
  done:
@@ -1119,6 +1178,8 @@ cli_rpc_controller_commit(clixon_handle h,
                 goto done;
             if (result != TR_SUCCESS)
                 goto ok;
+            if (transaction_print_skipped(h, tidstr) < 0)
+                goto done;
         }
         /* Interpret actions and no push as diff */
         if (actions_type_str2int(actions_type) != AT_NONE &&
@@ -1127,7 +1188,6 @@ cli_rpc_controller_commit(clixon_handle h,
                 goto done;
         }
     }
-    cligen_output(stderr, "OK\n");
  ok:
     retval = 0;
  done:
@@ -1229,8 +1289,8 @@ cli_connection_change(clixon_handle h,
         if (ret == 1){
             if (transaction_notification_poll(h, tidstr, &result) < 0)
                 goto done;
-            if (result != TR_SUCCESS)
-                cligen_output(stderr, "OK\n");
+            if (transaction_print_skipped(h, tidstr) < 0)
+                goto done;
         }
     }
     retval = 0;
@@ -1309,6 +1369,9 @@ show_connections_pretty(clixon_handle h,
         }
         cligen_output(stdout, "%-24s",  name);
         state = xml_find_body(xc, "conn-state");
+        if (xml_find_body(xc, "enabled") != NULL &&
+            strcmp(xml_find_body(xc, "enabled"), "true") != 0)
+            state = "DISABLED";
         cligen_output(stdout, "%-11s",  state?state:"");
         if ((timestamp = xml_find_body(xc, "stable-timestamp")) != NULL){
             /* Remove 6 us digits */
@@ -1461,7 +1524,7 @@ cli_show_connections(clixon_handle h,
     else{
         /* Avoid including moint-point which triggers a lot of extra traffic */
         if (clicon_rpc_get(h,
-                           "co:devices/co:device/co:name | co:devices/co:device/co:conn-state | co:devices/co:device/co:stable-timestamp | co:devices/co:device/co:logmsg",
+                           "co:devices/co:device/co:name | co:devices/co:device/co:enabled | co:devices/co:device/co:conn-state | co:devices/co:device/co:stable-timestamp | co:devices/co:device/co:logmsg",
                            nsc, CONTENT_ALL, -1, "explicit", &xn) < 0)
             goto done;
     }
@@ -1896,6 +1959,8 @@ compare_device_config_type(clixon_handle      h,
             if (result != TR_SUCCESS)
                 goto done;
         }
+        if (transaction_print_skipped(h, tidstr) < 0)
+            goto done;
     }
     if ((cb = cbuf_new()) == NULL){
         clixon_err(OE_PLUGIN, errno, "cbuf_new");
@@ -2109,9 +2174,7 @@ check_device_db(clixon_handle h,
     if (compare_device_config_type(h, cvv, argv, DT_RUNNING, DT_TRANSIENT, cbdiff) < 0)
         goto done;
     if (strlen(cbuf_get(cbdiff)))
-        cligen_output(stdout, "device out-of-sync\n");
-    else
-        cligen_output(stdout, "OK\n");
+        cligen_output(stdout, "Error: device out-of-sync\n");
     retval = 0;
  done:
     if (cbdiff)
@@ -2958,8 +3021,8 @@ cli_device_rpc_template(clixon_handle h,
     if (ret == 1){
         if (transaction_notification_poll(h, tidstr, &result) < 0)
             goto done;
-        if (result != TR_SUCCESS)
-            cligen_output(stderr, "OK\n");
+        if (transaction_print_skipped(h, tidstr) < 0)
+            goto done;
     }
     retval = 0;
  done:
@@ -3269,8 +3332,8 @@ cli_show_device_state(clixon_handle h,
     if (ret == 1){
         if (transaction_notification_poll(h, tidstr, &result) < 0)
             goto done;
-        if (result != TR_SUCCESS)
-            cligen_output(stderr, "OK\n");
+        if (transaction_print_skipped(h, tidstr) < 0)
+                goto done;
     }
     retval = 0;
  done: